// types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.5';
  };
  public: {
    Tables: {
// ...
      conversation_participants: {
        Row: {
          conversation_id: string | null;
          created_at: string;
          id: string;
          user_id: string | null;
        };
        Insert: {
          conversation_id?: string | null;
          created_at?: string;
          id?: string;
          user_id?: string | null;
        };
        Update: {
          conversation_id?: string | null;
          created_at?: string;
          id?: string;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: 'conversation_participants_conversation_id_fkey';
            columns: ['conversation_id'];
            isOneToOne: false;
            referencedRelation: 'conversations';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'conversation_participants_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          },
        ];
      };
      conversations: {
        Row: {
          created_at: string;
          id: string;
        };
        Insert: {
          created_at?: string;
          id?: string;
        };
        Update: {
          created_at?: string;
          id?: string;
        };
        Relationships: [];
      };
// ...
      messages: {
        Row: {
          content: string | null;
          conversation_id: string | null;
          created_at: string;
          id: string;
          image_path: string | null;
          sender_id: string | null;
        };
        Insert: {
          content?: string | null;
          conversation_id?: string | null;
          created_at?: string;
          id?: string;
          image_path?: string | null;
          sender_id?: string | null;
        };
        Update: {
          content?: string | null;
          conversation_id?: string | null;
          created_at?: string;
          id?: string;
          image_path?: string | null;
          sender_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: 'messages_conversation_id_fkey';
            columns: ['conversation_id'];
            isOneToOne: false;
            referencedRelation: 'conversations';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      requesting_user_id: { Args: never; Returns: string };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};